refactor: manager에서 generator로 이름 변경
