+++ src/main/java/koreatech/in/service/OwnerServiceImpl.java
@@ -30,7 +30,7 @@ import koreatech.in.util.SesMailSender;
 import koreatech.in.util.SlackNotiSender;
 import koreatech.in.util.StringRedisUtilObj;
 import koreatech.in.util.StringRedisUtilStr;
-import koreatech.in.util.jwt.TemporaryAccessJwtManager;
+import koreatech.in.util.jwt.TemporaryAccessJwtGenerator;
 import org.apache.velocity.app.VelocityEngine;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -57,7 +57,7 @@ public class OwnerServiceImpl implements OwnerService {
     private JwtValidator jwtValidator;
 
     @Autowired
-    private TemporaryAccessJwtManager temporaryAccessJwtGenerator;
+    private TemporaryAccessJwtGenerator temporaryAccessJwtGenerator;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
