+++ src/main/java/koreatech/in/service/UserServiceImpl.java
@@ -47,8 +47,8 @@ import koreatech.in.repository.user.StudentMapper;
 import koreatech.in.repository.user.UserMapper;
 import koreatech.in.util.SesMailSender;
 import koreatech.in.util.SlackNotiSender;
-import koreatech.in.util.jwt.UserAccessJwtManager;
-import koreatech.in.util.jwt.UserRefreshJwtManager;
+import koreatech.in.util.jwt.UserAccessJwtGenerator;
+import koreatech.in.util.jwt.UserRefreshJwtGenerator;
 import org.apache.velocity.app.VelocityEngine;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -84,10 +84,10 @@ public class UserServiceImpl implements UserService, UserDetailsService {
     private RefreshJwtValidator refreshJwtValidator;
 
     @Autowired
-    private UserAccessJwtManager userAccessJwtGenerator;
+    private UserAccessJwtGenerator userAccessJwtGenerator;
 
     @Autowired
-    private UserRefreshJwtManager userRefreshJwtGenerator;
+    private UserRefreshJwtGenerator userRefreshJwtGenerator;
 
     @Autowired
     private SlackNotiSender slackNotiSender;
